ovs-vsctl bash completion
This commit adds preliminary bash completion support for ovs-vsctl with
context-sensitive completion of arguments.

In order to do this in an extensible way, --commands and --options
arguments have been added to ovs-vsctl.  They derive data from the
internal option and command tables, so adding new commands or options
will result in the bash completion detecting the new commands or
options.

Signed-off-by: Peter Amidon <peter@picnicpark.org>