Fix reprinting prompt in NEW-* completion functions
Previously the NEW-* functions attempted to directly print out $PS1
which does not always work as many escape characters that can be present
in $PS1 are not handled by printf.  This commit changes it tuo use a
dummy function which currently invokes "bash -i" to get a copy of the
prompt.