Make _ovs_vsctl_bashcomp_globalopt work with =ARG options
The _ovs_vsctl_bashcomp_globalopt function would previously erroneously
return the ARG part of the option if an argument with the = in place was
completed; for example, '--db' would complete to '--db= ', but '--db='
would complete to '--db=ARG '.  This makes '--db=' properly complete to
'--db=', and removes the space after the equals for arguments with a
value.